+++ app/src/main/java/com/umc/upstyle/data/model/ClothPreview.kt
@@ -1,14 +1,14 @@
 package com.umc.upstyle.data.model
 
 data class ClothPreview(
-    val id: Long,
-    val kindId: Long,
+    val id: Int,
+    val kindId: Int,
     val kindName: String,
-    val categoryId: Long,
+    val categoryId: Int,
     val categoryName: String,
-    val fitId: Long,
+    val fitId: Int,
     val fitName: String,
-    val colorId: Long,
+    val colorId: Int,
     val colorName: String,
     val additionalInfo: String?,
     val ootd: Ootd? = null  // ✅ Ootd를 nullable로 변경
